chore: split basic.test.ts into smaller ones
