+++ test/resource-embedding.test.ts
@@ -37,6 +37,48 @@ test('embedded select', async () => {
             Object {
               "channel_id": 1,
               "data": null,
+              "id": 5,
+              "message": "foo",
+              "username": "supabot",
+            },
+            Object {
+              "channel_id": 1,
+              "data": null,
+              "id": 6,
+              "message": "foo",
+              "username": "supabot",
+            },
+            Object {
+              "channel_id": 1,
+              "data": null,
+              "id": 7,
+              "message": "foo",
+              "username": "supabot",
+            },
+            Object {
+              "channel_id": 1,
+              "data": null,
+              "id": 8,
+              "message": "foo",
+              "username": "supabot",
+            },
+            Object {
+              "channel_id": 1,
+              "data": null,
+              "id": 9,
+              "message": "foo",
+              "username": "supabot",
+            },
+            Object {
+              "channel_id": 1,
+              "data": null,
+              "id": 10,
+              "message": "foo",
+              "username": "supabot",
+            },
+            Object {
+              "channel_id": 1,
+              "data": null,
               "id": 11,
               "message": "foo",
               "username": "supabot",
@@ -160,6 +202,48 @@ describe('embedded filters', () => {
                 "data": null,
                 "id": 1,
                 "message": "Hello World 👋",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 5,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 6,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 7,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 8,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 9,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 10,
+                "message": "foo",
                 "username": "supabot",
               },
               Object {
@@ -454,6 +538,48 @@ describe('embedded transforms', () => {
               Object {
                 "channel_id": 1,
                 "data": null,
+                "id": 7,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 8,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 9,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 10,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 11,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 12,
+                "message": "test1",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
                 "id": 13,
                 "message": "test1",
                 "username": "supabot",
@@ -503,15 +629,15 @@ describe('embedded transforms', () => {
               Object {
                 "channel_id": 1,
                 "data": null,
-                "id": 11,
+                "id": 5,
                 "message": "foo",
                 "username": "supabot",
               },
               Object {
                 "channel_id": 1,
                 "data": null,
-                "id": 12,
-                "message": "test1",
+                "id": 6,
+                "message": "foo",
                 "username": "supabot",
               },
             ],
@@ -595,6 +721,48 @@ describe('embedded transforms', () => {
               Object {
                 "channel_id": 1,
                 "data": null,
+                "id": 7,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 8,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 9,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 10,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 11,
+                "message": "foo",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
+                "id": 12,
+                "message": "test1",
+                "username": "supabot",
+              },
+              Object {
+                "channel_id": 1,
+                "data": null,
                 "id": 13,
                 "message": "test1",
                 "username": "supabot",
@@ -644,15 +812,15 @@ describe('embedded transforms', () => {
               Object {
                 "channel_id": 1,
                 "data": null,
-                "id": 11,
+                "id": 5,
                 "message": "foo",
                 "username": "supabot",
               },
               Object {
                 "channel_id": 1,
                 "data": null,
-                "id": 12,
-                "message": "test1",
+                "id": 6,
+                "message": "foo",
                 "username": "supabot",
               },
             ],
